Features/update client (#4)
* Fix parameter type of BaseRepository.GetByIdAsync

* Update client handler is created

* Update method is added into client controller

+++ POS.Infrastructure/Repositories/BaseRepository.cs
@@ -22,7 +22,7 @@ namespace POS.Infrastructure.Repositories
             return await _context.Set<T>().ToListAsync();
         }
 
-        public virtual async Task<T?> GetByIdAsync(long id)
+        public virtual async Task<T?> GetByIdAsync(uint id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
